Created a config for the entire cli
