+++ synth/src/cli/telemetry.rs
@@ -5,8 +5,11 @@ use std::collections::HashMap;
 use std::path::PathBuf;
 use std::future::Future;
 use std::error::Error;
+use uuid::Uuid;
 
-use crate::utils::{version, META_OS};
+use crate::utils::META_OS;
+use crate::version::version;
+use crate::cli::config;
 
 use super::{Args, TelemetryCommand};
 
@@ -54,62 +57,20 @@ struct TelemetryConfig {
 }
 
 impl TelemetryConfig {
-    pub fn initialise() -> Self {
-        Self::from_file().unwrap_or_else(|_| Self::new())
-    }
-
-    fn new() -> Self {
-        Self {
-            uuid: uuid::Uuid::new_v4().to_hyphenated().to_string(),
-        }
-    }
-
-    fn from_file() -> Result<Self> {
-        let file_contents = std::fs::read_to_string(Self::file_path()?)?;
-        let tc = serde_json::from_str(&file_contents)?;
-        Ok(tc)
-    }
-
-    fn synth_config_dir() -> Result<PathBuf> {
-        let synth_config_dir = dirs::config_dir().ok_or_else(|| {
-            anyhow!(
-                "Could not find a configuration directory. Your operating system may not be supported."
-            )
-        })?;
-        Ok(synth_config_dir.join("synth"))
-    }
-
-    fn file_path() -> Result<PathBuf> {
-        Ok(Self::synth_config_dir()?.join("config.json"))
-    }
-
     fn enable_telemetry() -> Result<()> {
-	let config_dir = Self::synth_config_dir()?;
-        if !config_dir.exists() {
-            std::fs::create_dir_all(&config_dir)
-		.with_context(|| anyhow!("Could not create the directory: {}", config_dir.display()))?;
+        // Initialise the `uuid` if it hasn't been initialised yet.
+        if config::get_uuid().is_none() {
+            config::set_uuid(Uuid::new_v4().to_hyphenated().to_string());
         }
-        if !Self::is_enabled() {
-            let mut config_file_path = std::fs::OpenOptions::new()
-                .write(true)
-                .create(true)
-                .open(Self::file_path()?)
-                .map_err(|e| anyhow!("There was an issue enabling telemetry: {}", e))?;
-            serde_json::to_writer_pretty(&mut config_file_path, &TelemetryConfig::new())?;
-        }
-        Ok(())
+        Ok(config::set_telemetry_enabled(true))
     }
 
     fn disable_telemetry() -> Result<()> {
-        if Self::is_enabled() {
-            std::fs::remove_file(Self::file_path()?)
-                .map_err(|e| anyhow!("There was an issue disabling telemetry: {}", e))?;
-        }
-        Ok(())
+        Ok(config::set_telemetry_enabled(false))
     }
 
     fn is_enabled() -> bool {
-        Self::file_path().map(|path| path.exists()).unwrap_or(false)
+        config::get_telemetry_enabled().unwrap_or(false)
     }
 }
